Should have simpler, more meaningful header names
The main modification was renaming types to box.
