+++ walls-of-doom/box.h
@@ -1,5 +1,5 @@
-#ifndef TYPES_H
-#define TYPES_H
+#ifndef BOX_H
+#define BOX_H
 
 typedef struct BoundingBox {
   int min_x;
