+++ walls-of-doom/perk.h
@@ -1,15 +1,13 @@
-#ifndef PERKS_H
-#define PERKS_H
+#ifndef PERK_H
+#define PERK_H
 
 #define PERK_INTERVAL_IN_SECONDS 30
 #define PERK_INTERVAL_IN_FRAMES PERK_INTERVAL_IN_SECONDS *FPS
 
-#define PERK_DURATION_ON_SCREEN_IN_SECONDS 15
-#define PERK_DURATION_ON_PLAYER_IN_SECONDS 10
-#define PERK_DURATION_ON_SCREEN_IN_FRAMES                                      \
-  PERK_DURATION_ON_SCREEN_IN_SECONDS *FPS
-#define PERK_DURATION_ON_PLAYER_IN_FRAMES                                      \
-  PERK_DURATION_ON_PLAYER_IN_SECONDS *FPS
+#define PERK_SCREEN_DURATION_IN_SECONDS 15
+#define PERK_PLAYER_DURATION_IN_SECONDS 10
+#define PERK_SCREEN_DURATION_IN_FRAMES PERK_SCREEN_DURATION_IN_SECONDS *FPS
+#define PERK_PLAYER_DURATION_IN_FRAMES PERK_PLAYER_DURATION_IN_SECONDS *FPS
 
 typedef enum Perk {
   PERK_POWER_INVINCIBILITY,
