+++ walls-of-doom/physics.h
@@ -1,11 +1,11 @@
 #ifndef PHYSICS_H
 #define PHYSICS_H
 
+#include "box.h"
 #include "command.h"
 #include "game.h"
 #include "platform.h"
 #include "player.h"
-#include "types.h"
 
 #include "stdlib.h"
