+++ walls-of-doom/player.h
@@ -1,7 +1,7 @@
 #ifndef PLAYER_H
 #define PLAYER_H
 
-#include "perks.h"
+#include "perk.h"
 
 typedef struct Player {
